Moved model classes in model package
