+++ richtextfx-demos/src/main/java/org/fxmisc/richtext/demo/XMLEditor.java
@@ -8,8 +8,8 @@ import java.util.regex.Pattern;
 import org.fxmisc.flowless.VirtualizedScrollPane;
 import org.fxmisc.richtext.CodeArea;
 import org.fxmisc.richtext.LineNumberFactory;
-import org.fxmisc.richtext.StyleSpans;
-import org.fxmisc.richtext.StyleSpansBuilder;
+import org.fxmisc.richtext.model.StyleSpans;
+import org.fxmisc.richtext.model.StyleSpansBuilder;
 
 import javafx.application.Application;
 import javafx.scene.Scene;
