+++ richtextfx/src/main/java/org/fxmisc/richtext/ClipboardActions.java
@@ -14,6 +14,8 @@ import javafx.scene.input.Clipboard;
 import javafx.scene.input.ClipboardContent;
 import javafx.scene.input.DataFormat;
 
+import org.fxmisc.richtext.model.ReadOnlyStyledDocument;
+import org.fxmisc.richtext.model.StyledDocument;
 import org.reactfx.util.Tuple2;
 
 /**
