+++ richtextfx/src/main/java/org/fxmisc/richtext/InlineCssTextArea.java
@@ -1,6 +1,9 @@
 package org.fxmisc.richtext;
 
 
+import org.fxmisc.richtext.model.EditableStyledDocument;
+import org.fxmisc.richtext.model.EditableStyledDocumentImpl;
+
 import javafx.scene.text.TextFlow;
 
 /**
