+++ richtextfx/src/main/java/org/fxmisc/richtext/StyleClassedTextArea.java
@@ -3,6 +3,9 @@ package org.fxmisc.richtext;
 import java.util.Collection;
 import java.util.Collections;
 
+import org.fxmisc.richtext.model.EditableStyledDocument;
+import org.fxmisc.richtext.model.EditableStyledDocumentImpl;
+
 /**
  * Text area that uses style classes to define style of text segments and paragraph segments.
  */
