+++ richtextfx/src/main/java/org/fxmisc/richtext/StyledTextArea.java
@@ -53,6 +53,16 @@ import org.fxmisc.flowless.VirtualFlowHit;
 import org.fxmisc.flowless.Virtualized;
 import org.fxmisc.flowless.VirtualizedScrollPane;
 import org.fxmisc.richtext.CssProperties.EditableProperty;
+import org.fxmisc.richtext.model.EditableStyledDocument;
+import org.fxmisc.richtext.model.EditableStyledDocumentImpl;
+import org.fxmisc.richtext.model.Paragraph;
+import org.fxmisc.richtext.model.PlainTextChange;
+import org.fxmisc.richtext.model.RichTextChange;
+import org.fxmisc.richtext.model.StyleSpans;
+import org.fxmisc.richtext.model.StyledDocument;
+import org.fxmisc.richtext.model.StyledTextAreaModel;
+import org.fxmisc.richtext.model.TwoDimensional;
+import org.fxmisc.richtext.model.TwoLevelNavigator;
 import org.fxmisc.undo.UndoManager;
 import org.fxmisc.undo.UndoManagerFactory;
 import org.reactfx.EventStream;
