+++ richtextfx/src/main/java/org/fxmisc/richtext/TextEditingArea.java
@@ -4,6 +4,10 @@ import javafx.beans.value.ObservableValue;
 import javafx.collections.ObservableList;
 import javafx.scene.control.IndexRange;
 
+import org.fxmisc.richtext.model.Paragraph;
+import org.fxmisc.richtext.model.PlainTextChange;
+import org.fxmisc.richtext.model.RichTextChange;
+import org.fxmisc.richtext.model.StyledDocument;
 import org.reactfx.EventStream;
 
 /**
