+++ richtextfx/src/main/java/org/fxmisc/richtext/model/StyleSpan.java
@@ -1,4 +1,4 @@
-package org.fxmisc.richtext;
+package org.fxmisc.richtext.model;
 
 import java.util.Objects;
