+++ richtextfx/src/test/java/org/fxmisc/richtext/TwoLevelNavigatorTest.java
@@ -1,9 +1,10 @@
 package org.fxmisc.richtext;
 
-import static org.fxmisc.richtext.TwoDimensional.Bias.*;
+import static org.fxmisc.richtext.model.TwoDimensional.Bias.*;
 import static org.junit.Assert.*;
 
-import org.fxmisc.richtext.TwoDimensional.Position;
+import org.fxmisc.richtext.model.TwoLevelNavigator;
+import org.fxmisc.richtext.model.TwoDimensional.Position;
 import org.junit.Test;
 
 public class TwoLevelNavigatorTest {
